Ajuste de variável pra double

+++ EncoderLED/1_global.ino
@@ -1,10 +1,10 @@
 //Encoder
 #include <Encoder.h>
 Encoder myEnc(D1, D2);
-long encoderOldPosition  = -999;
-long encoderNewPosition;
+double encoderOldPosition  = -999;
+double encoderNewPosition;
 boolean isEncoderButtonPressed = false;
-long lastUpdateEncoderMillis = 0;
+double lastUpdateEncoderMillis = 0;
 
 //Led
 #define ledR D7
@@ -13,6 +13,7 @@ long lastUpdateEncoderMillis = 0;
 #define hueFactor 1023 //esse valor é a multiplicação do valor da matiz, que retorna de 0 a 1.
 #define brightnessFactor 0.2; //valor de decaimento fixo (subtraído) ao apertar o botão
 
+//Cores
 double r = 0;
 double g = 0;
 double b = 0;
